+++ tyrannosaurus/parser.py
@@ -53,6 +53,7 @@ class LiteralParser:
             "Project": self.project.capitalize(),
             "PROJECT": self.project.upper(),
             "pkg": self.pkg,
+            "Pkg": self.pkg.upper(),
             "user": "<<TODO:user>>" if self.user is None else self.user,
             "authors": self._pretty(self.authors),
             "authors.list": self._list(self.authors),